+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- use Deck's assignment/permission classes to resolve group/circle users and enforce ACL 1.1.1
 - send notifications on start/end/edit/reminder (after 1h) 1.1.0
 - /js/ wasn't pushed in git & fixed timesheet table on sidebar details 1.0.2
 - Fix for production 1.0.1